Responsive + likes + json

+++ WEB-INF/java/src/dao/SubscriberDAO.java
@@ -10,13 +10,22 @@ public class SubscriberDAO {
     }
 
     // Ajouter un abonnement
-    public boolean addSubscriber(int userId, int threadId) throws SQLException {
-        try(Connection connection = ds.getConnection()) {
-            String sql = "INSERT INTO subscriber (usr_id, thread_id) VALUES (?, ?)";
+    public void addSubscriber(int userId, int threadId, int adminId) throws SQLException {
+        try (Connection connection = ds.getConnection()) {
+            String sql = "SELECT admin_id FROM thread WHERE id = ?";
             try (PreparedStatement stmt = connection.prepareStatement(sql)) {
-                stmt.setInt(1, userId);
-                stmt.setInt(2, threadId);
-                return stmt.executeUpdate() > 0;
+                stmt.setInt(1, threadId);
+                ResultSet rs = stmt.executeQuery();
+                if (rs.next() && rs.getInt("admin_id") == adminId) {
+                    String sql2 = "INSERT INTO subscriber (usr_id, thread_id) VALUES (?, ?)";
+                    try (PreparedStatement stmt2 = connection.prepareStatement(sql2)) {
+                        stmt2.setInt(1, userId);
+                        stmt2.setInt(2, threadId);
+                        stmt2.executeUpdate();
+                    }catch (SQLException e){
+                        System.out.println(e.getMessage());
+                    }
+                }
             }
         }
 
@@ -36,13 +45,22 @@ public class SubscriberDAO {
     }
 
     // Supprimer un abonnement
-    public boolean removeSubscriber(int userId, int threadId) throws SQLException {
+    public void removeSubscriber(int userId, int threadId, int admin_id) throws SQLException {
         try(Connection connection = ds.getConnection()) {
-            String sql = "DELETE FROM subscriber WHERE usr_id = ? AND thread_id = ?";
+            String sql = "SELECT admin_id FROM thread WHERE id = ?";
             try (PreparedStatement stmt = connection.prepareStatement(sql)) {
-                stmt.setInt(1, userId);
-                stmt.setInt(2, threadId);
-                return stmt.executeUpdate() > 0;
+                stmt.setInt(1, threadId);
+                ResultSet rs = stmt.executeQuery();
+                if (rs.next() && rs.getInt("admin_id") == admin_id) {
+                    String sql2 = "DELETE FROM subscriber WHERE usr_id = ? AND thread_id = ?";
+                    try (PreparedStatement stmt2 = connection.prepareStatement(sql2)) {
+                        stmt2.setInt(1, userId);
+                        stmt2.setInt(2, threadId);
+                        stmt2.executeUpdate();
+                    }catch (SQLException e){
+                        System.out.println(e.getMessage());
+                    }
+                }
             }
         }
     }
